refactor: extract chart color mappings to config file
Move the hardcoded CHART_COLOR_MAPPING from formatCharts.tsx into a
dedicated config file at app/config/chartColorMappings.ts. This
resolves the existing TODO and makes it easier to extend or generate
color mappings from dataset metadata in the future.

New file: app/config/chartColorMappings.ts
Modified: app/utils/formatCharts.tsx
