+++ app/utils/formatCharts.tsx
@@ -1,4 +1,5 @@
 import getChartColors from "./ChartColors";
+import CHART_COLOR_MAPPING from "@/app/config/chartColorMappings";
 
 interface InputData {
   [key: string]: unknown | unknown;
@@ -6,11 +7,6 @@ interface InputData {
 
 interface ChartData {
   [key: string]: unknown;
-}
-
-interface ColorMapEntry {
-  value: string;
-  color: string;
 }
 
 interface ChartSeries {
@@ -28,54 +24,6 @@ interface ChartSeries {
  * @param yAxis The key to use for the y-axis (required for scatter charts).
  * @returns An object containing the transformed `data` and `series` arrays.
  */
-
-//TODO: Generate this from the DATASET_CARDS fixture or move to config
-const CHART_COLOR_MAPPING: Record<string, ColorMapEntry[]> = {
-  "land_cover_type": [
-    { value: "Tree cover", color: "#246E24" },
-    { value: "Short vegetation", color: "#B9B91E" },
-    { value: "Wetland – short vegetation", color: "#74D6B4" },
-    { value: "Bare and sparse vegetation", color: "#FEFECC" },
-    { value: "Water", color: "#6BAED6" },
-    { value: "Snow/ice", color: "#ACD1E8" },
-    { value: "Cropland", color: "#fff183" },
-    { value: "Cultivated grasslands", color: "#FFCD73" },
-    { value: "Built-up", color: "#e8765d" },
-  ],
-  "land_type": [
-    {value: "Natural forests", color: "#246E24" },
-    { value: "Natural peat forests", color: "#093D09" },
-    { value: "Natural peat short vegetation", color: "#99991A" },
-    { value: "Mangroves", color: "#06A285" },
-    { value: "Wet natural forests", color: "#589558" },
-    { value: "Wet natural short vegetation", color: "#DBDB7B" },
-    { value: "Natural short vegetation", color: "#B9B91E" },
-    { value: "Natural water", color: "#6BAED6" },
-    { value: "Bare", color: "#FEFECC" },
-    { value: "Snow", color: "#ACD1E8" },
-    { value: "Crop", color: "#D3D3D3" },
-    { value: "Built", color: "#D3D3D3" },
-    { value: "Non-natural tree cover", color: "#D3D3D3" },
-    { value: "Non-natural short vegetation", color: "#D3D3D3" },
-    { value: "Wet non-natural tree cover", color: "#D3D3D3" },
-    { value: "Non-natural peat tree cover", color: "#D3D3D3" },
-    { value: "Wet non-natural short vegetation", color: "#D3D3D3" },
-    { value: "Non-natural peat short vegetation", color: "#D3D3D3" },
-    { value: "Non-natural water", color: "#D3D3D3" },
-    { value: "Non-natural bare", color: "#D3D3D3" },
-    { value: "Other", color: "#D3D3D3" }],
-  "driver": [
-    { value: "Logging", color: "#52A44E"},
-    { value: "Shifting cultivation", color: "#E9D700"},
-    { value: "Wildfire", color: "#885128"},
-    { value: "Other natural disturbances", color: "#3B209A"},
-    { value: "Settlements & Infrastructure", color: "#A354A0"},
-    { value: "Hard commodities", color: "#E58074"},
-    { value: "Permanent agriculture", color: "#E39D29"},
-    { value: "Unknown", color: "#D3D3D3"},
-  ]
-
-}
 export default function formatChartData(
   data: InputData[] | unknown,
   type:
